Set mediator "zoom out" in store
Use Redux Middleware to provide a place to transparently
inject a function to be called on zoomOut.

Probably makes most sense to do it like this in the constructor.

We can probably add additional callback functions as needed.
E.g. we could add a separate one for zoomOut type behaviour in
stream graph.